fix test suite to demonstrate working static synapse
